add mail before 05.04

+++ app/Http/Controllers/ArticleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Article;
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -22,6 +23,7 @@ class ArticleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', [self::class]);
         return view('article.create');
     }
 
@@ -58,6 +60,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        $this->authorize('update', [self::class, $article]);
         return view('article.edit', ['article'=>$article]);
 
     }
@@ -85,6 +88,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('delete', [self::class, $article]);
         $article->delete();
         return redirect()->route('article.index');
     }
